backend/service/requirements.txt
Flask==3.1.0
Werkzeug==3.1.0
mysql-connector==2.2.9
flask-cors
flask[async]>=2.0.0
requests
